real-time-data-platform/airflow/requirements.txt
confluent-kafka==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jsonschema==4.23.0
